Added isolated Cheatnet testing
