+++ .gitignore
@@ -3,3 +3,4 @@ pyproject.toml
 .DS_Store
 .tool-versions
 target
+!crates/cheatnet/tests/contracts/target
