+++ crates/cheatnet/src/cheatcodes/deploy.rs
@@ -39,7 +39,7 @@ impl CheatnetState {
         let salt = ContractAddressSalt::default();
 
         let contract_class = blockifier_state
-            .get_compiled_contract_class(&class_hash)
+            .get_compiled_contract_class(class_hash)
             .map_err::<EnhancedHintError, _>(From::from)?;
         if contract_class.constructor_selector().is_none() && !calldata.is_empty() {
             return Err(CheatcodeError::Recoverable(vec![felt_from_short_string(
@@ -49,7 +49,7 @@ impl CheatnetState {
 
         let execute_calldata = create_execute_calldata(
             calldata,
-            &class_hash,
+            class_hash,
             &account_address,
             &entry_point_selector,
             &salt,
